test(IntegrationTest): 테스트 상속 구조 작성
