+++ src/test/java/subway/StationIntegrationTest.java
@@ -1,15 +1,8 @@
 package subway;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.jdbc.core.JdbcTemplate;
 import subway.dto.StationCreateRequest;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -17,25 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.springframework.http.HttpStatus.CREATED;
 import static subway.steps.StationSteps.역_생성_요청;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class StationIntegrationTest {
-    @LocalServerPort
-    int port;
-
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-
-    @BeforeEach
-    public void setUp() {
-        RestAssured.port = port;
-    }
-
-    @AfterEach
-    void tearDown() {
-        jdbcTemplate.update("TRUNCATE TABLE station");
-        jdbcTemplate.update("TRUNCATE TABLE line");
-        jdbcTemplate.update("TRUNCATE TABLE section");
-    }
+class StationIntegrationTest extends IntegrationTest {
 
     @Test
     void createStationTest() {
